hv: lib: split atomic operation to atomic.h
split atomic related operation to file atomic.h

Acked-by: Eddie Dong <eddie.dong@intel.com>
